bugfix & enhancement: remade tables as Mui components & fixed no group section bug (#300 & #550)
