--- src/components/pages/account-page/components/group-tab/components/table/mobile-requests-table/MobileRequestsTable.module.scss
@@ -1,77 +1,0 @@
-@import "../../../../../../../../styles/partials/theme";
-
-.table {
-  margin-bottom: 58px;
-}
-
-.table-container {
-  display: flex;
-  width: 100%;
-  height: 60px;
-  padding: 12px 10px;
-  gap: 12px;
-  background-color: $background-dark-10;
-  border-top: 1px solid $gray-50;
-  border-left: 1px solid $gray-50;
-  border-right: 1px solid $gray-50;
-  border-bottom: 0.5px solid $gray-50;
-  img {
-    width: 36px;
-    height: 36px;
-    border-radius: 100%;
-  }
-}
-
-.table-container-start{
-  @extend .table-container;
-  border-radius: 6px 6px 0 0;
-}
-
-.table-container-end{
-  @extend .table-container;
-  border-radius: 0 0 6px 6px;
-  border-bottom: 1px solid $gray-50;;
-}
-
-.table-container-one{
-  @extend .table-container;
-  border-radius: 6px;
-  border-bottom: 1px solid $gray-50;
-}
-
-.user-info{
-  width: 52%;
-  .full-name {
-    font-weight: 700;
-    font-size: 14px;
-    line-height: 140%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-  }
-
-  .email {
-    font-weight: 400;
-    font-size: 11px;
-    line-height: 160%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-  }
-}
-
-.button-left{
-  height: 40px;
-}
-
-.button-right{
-  @extend .button-left;
-  background-color: $background-dark-35;
-}
-
-.side-buttons {
-  display: flex;
-  gap: 5px;
-  margin-left: auto;
-  align-items: center;
-}
