--- src/components/pages/account-page/components/group-tab/components/table/mobile-requests-table/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MobileRequestsTable';
