+++ src/components/pages/account-page/components/group-tab/components/table/requests-table/RequestTable.styles.ts
@@ -1,0 +1,8 @@
+import { SxProps } from '@mui/material/styles';
+import { Theme } from '@mui/system';
+
+export const fixSized: SxProps<Theme> = { width: 'fit-content' };
+export const lastColumn: SxProps<Theme> = {
+  justifyContent: 'flex-end',
+  gap: '8px',
+};
