--- src/components/pages/account-page/components/group-tab/components/table/requests-table/RequestsTable.module.scss
@@ -1,76 +1,0 @@
-@import "../../../../../../../../styles/partials/theme";
-
-.table {
-  margin-bottom: 58px;
-}
-
-.table-container {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 72px;
-  padding: 12px 16px;
-  gap: 26px;
-  background-color: $background-dark-10;
-  border-top: 1px solid $gray-50;
-  border-left: 1px solid $gray-50;
-  border-right: 1px solid $gray-50;
-  border-bottom: 0.5px solid $gray-50;
-}
-
-.table-container-start{
-  @extend .table-container;
-  border-radius: 6px 6px 0 0;
-}
-
-.table-container-end{
-  @extend .table-container;
-  border-radius: 0 0 6px 6px;
-  border-bottom: 1px solid $gray-50;
-}
-
-.table-container-one{
-  @extend .table-container;
-  border-radius: 6px;
-  border-bottom: 1px solid $gray-50;
-}
-
-.user-info{
-  display: flex;
-  align-items: center;
-  gap: 3%;
-  width: 45%;
-  height: 100%;
-  .full-name{
-    display: flex;
-    align-items: center;
-    width: 60%;
-    height: 100%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  .img{
-    height: 48px;
-    width: 48px;
-    border-radius: 100%;
-  }
-}
-
-.other-content{
-  width: 60%;
-  display: flex;
-  justify-content: space-between;
-}
-
-.email {
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.side-buttons {
-  display: flex;
-  gap: 5%;
-}
