+++ src/components/pages/account-page/components/group-tab/components/table/student-table/components/EditingColumn.tsx
@@ -60,8 +60,6 @@ const EditingColumn: FC<EditingColumnProps> = ({ student, refetch }) => {
       );
     }
   };
-
-  if (user.group?.role) user.group.role = UserGroupRole.MODERATOR;
   const handleChangeStatus = async () => {
     try {
       setChangePopupOpen(false);
